requirements.txt
google-generativeai>=0.8.0
jsonschema>=4.21.0
openpyxl>=3.1.2
python-dateutil>=2.9.0

